+++ app/Http/Controllers/MyJobController.php
@@ -72,6 +72,7 @@ class MyJobController extends Controller
     
     public function destroy(Job $myJob)
     {
+        $this->authorize('delete', $myJob);
         $myJob->delete();
 
         return redirect()->route('my-jobs.index')
